remove and rename Components
