+++ components/PageFooter.tsx
@@ -2,12 +2,10 @@ import React from 'react'
 
 import { ContainerBox } from './ContainerBox'
 
-const Footer: React.FC = () => {
+export const PageFooter: React.FC = () => {
   return (
     <footer className="bg-accent-1 border-t border-accent-2">
       <ContainerBox>&copy; Shoota Kumano all reserved</ContainerBox>
     </footer>
   )
 }
-
-export default Footer
