--- components/section-separator.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const SectionSeparator: React.FC = () => {
-  return <hr className="border-accent-2 mt-28 mb-24" />
-}
-
-export default SectionSeparator
